Artist/song order consistency; bug fixes
The genius link requires the artist to be first, and for some reason I was taking the song name first in the arguments, and also naming the text file [song name]-[artist] which was counter intuitive. Fixed a bug relating to trying to access the next item in an iterable when none existed. Fixed a bug that occured when multiple lines of italics occurs in a song.